Use database duration parser for consistency
