+++ backend/src/baserow/core/formula/utils/date.py
@@ -1,6 +1,6 @@
 import re
-from datetime import timedelta
-from typing import Optional
+
+from baserow.core.duration import parse_duration_string as parse_duration_string
 
 MOMENT_FORMAT_MAP = {
     "YYYY": "%Y",
@@ -34,32 +34,6 @@ RE_SUPPORTED_MOMENT_TOKEN = re.compile(
 RE_VALID_FORMAT = re.compile(
     "|".join(sorted(MOMENT_FORMAT_MAP.keys(), key=len, reverse=True)) + "|T"
 )
-
-DURATION_PATTERNS = [
-    (re.compile(r"^(\d+)\s+years?$", re.IGNORECASE), "days", 365),
-    (re.compile(r"^(\d+)\s+months?$", re.IGNORECASE), "days", 30),
-    (re.compile(r"^(\d+)\s+weeks?$", re.IGNORECASE), "weeks", 1),
-    (re.compile(r"^(\d+)\s+days?$", re.IGNORECASE), "days", 1),
-    (re.compile(r"^(\d+)\s+hours?$", re.IGNORECASE), "hours", 1),
-    (re.compile(r"^(\d+)\s+minutes?$", re.IGNORECASE), "minutes", 1),
-    (re.compile(r"^(\d+)\s+seconds?$", re.IGNORECASE), "seconds", 1),
-]
-
-
-def parse_duration_string(value: str) -> Optional[timedelta]:
-    """Parse a human-readable duration string into a timedelta, or None if invalid."""
-
-    if not isinstance(value, str):
-        return None
-
-    for pattern, unit, factor in DURATION_PATTERNS:
-        match = pattern.match(value.strip())
-        if match:
-            amount = int(match.group(1)) * factor
-            return timedelta(**{unit: amount})
-
-    return None
-
 
 def is_valid_datetime_format(value: str) -> bool:
     """Return True if the string contains only supported Moment.js format tokens."""
